Fix generate failing due to duplicate keys
